Refactored action implementation and added accessbility binding adapter
The refactoring of the action implementation was needed due to accessibility features. The current implementation always added a clickListener and longClickListener to the view. Due to that, tools like Talkback would always annonce that view with "Double tab to activate, double tab and hold to long press." With this new change, the dev should have more control over these default sentences.