Fixed bug in ArchTreeFragment which would initialise a toolbar for an invisible fragment
That would cause the actual visible fragment to have a not usuable toolbar (like close button etc). That was caused by the onCreate which included the refreshToolbar function but was missing a check if the fragment was actually visible at this point. This issue was actually a problem if the parent-activity has two fragments, the app gets a config change and both fragments are calling their onCreate again.