Allow fragments to create fragments
This commit also refactores the current fragment creation process. The builder artifact has it's own interface to determine if the class can dispatch new fragments. I've also added a new interface which is used for the builder process. That should reduce possible confusion why the same interface (implementation before) is also used within the component. Due to this change I made sure that there are no dependencies to the autotarget library anymore <-- autotarget will later use the builder artifact to choose which fragment creation mode to choose (auto vs class). Reason for this dependency shifting is to keep archtree as an utility library and autotarget, archknife, repolizer as consumer libraries.